Added some new models Updated inferrer and tests Added freezing to model object
Note that there are some problems in this commit. There is a new unit
test for extending the trainer history (for example, when restarting
training from a checkpoint). When re-training torch throws an error
saying that it was expecting a torch.FloatTensor but got a
torch.cuda.FloatTensor. I thought that this shouldn't occur since the
device is initialized in the trainer constructor and the models and data
are moved there at train time (specifically at construction time for the
model and in the call to train_epoch() for the data). This possibly
means that the current device mapping on checkpoint load is inadequate.

There is also a test for freezing and unfreezing the model that is not
extensive enough to show that freezing does hold the weight values
constant. This needs to be amended before merge.