Trying to add extra dataloaders for lr find
Actually it perhaps makes more sense for the LRFinder object itself to
hold datasets. It could be that by default they are none and it uses the
the trainers dataset, but being able to hold their own dataset
references means that we can load in custom datasets for learning rate
selection. The point here is really just to be able to use smaller
datasets. It doens't make sense to pick learning rates based on
'different' data, but some of the datasets are large enough that
iterating over them is very time consuming, to the point that finding
the learning rate takes nearly as long as training the network